chore(monitoring): split logAnalysisClient, logAnalysisManager from logAnalysisSevice #103
