+++ triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/monitoring/dto/MonitoringLogAnalysisRequestDto.java
@@ -1,6 +1,7 @@
 package com.triton.msa.triton_dashboard.monitoring.dto;
 
 public record MonitoringLogAnalysisRequestDto(
+        String esIndex,
         String provider,
         String model,
         String query
