+++ triton_dashboard/src/main/resources/application-dev.properties
@@ -6,3 +6,4 @@ elasticsearch.host=${ELASTICSEARCH_HOST:localhost}
 elasticsearch.port=${ELASTICSEARCH_PORT:9200}
 
 rag.server.url.monitoring=${RAG_SERVER_URL_MONITORING:http://rag-server-address/api/get-rag-response}
+monitoring.scheduler.rate=${MONITORING_SCHEDULER_RATE:180000}
